use renderer for frontend
