+++ src/helper/exercise-submission.ts
@@ -1,6 +1,6 @@
 import { v4 as uuidv4 } from 'uuid'
 
-interface ExerciseSubmissionData {
+export interface ExerciseSubmissionData {
   path: string
   entityId: number
   revisionId: number
