+++ src/helper/has-visible-content.ts
@@ -8,7 +8,8 @@ export function hasVisibleContent(content: FrontendContentNode[]): boolean {
 export function extractText(content: FrontendContentNode[]): string {
   return content
     .map((node) => {
-      if (node.type === 'math') return node.formula
+      if (node.type === 'math' || node.type === 'inline-math')
+        return node.formula
       if (Object.hasOwn(node, 'text')) {
         return node.text
       } else {
